+++ libs/base_core/src/ui/front/mod.rs
@@ -1,0 +1,14 @@
+// devela_base_core::ui::front
+//
+//! UI frontends.
+//
+
+pub mod term;
+
+crate::structural_mods! { // _pub_mods
+    _pub_mods {
+        pub use super::{
+            term::_all::*,
+        };
+    }
+}