+++ src/ui/front/term/mod.rs
@@ -3,12 +3,14 @@
 //! Terminal functionality.
 //
 
-mod size;
 mod ansi;
 
 crate::structural_mods! { // _mods
     _mods {
-        pub use super::size::*;
-        pub use super::ansi::*;
+        pub use super::ansi::_all::*;
+
+        // re-exports
+        #[doc(inline)]
+        pub use devela_base_core::ui::front::term::{Ansi, AnsiColor3b, AnsiColor8b, TermSize};
     }
 }